Finalize most of Qt4 support
The output of moc-ng is now mostly equal to Qt4's moc tool, except for
some comments and version information in the header of the output file.

Property type support is lacking, and probably some other Qt4 specific
functionality is lacking, which will be added as soon as the need for it
arises.

moc-ng-qt4 is certainly not 100% feature complete compared to Qt4's moc
tool.